+++ src/mongo/s/commands/cluster_hello_cmd.cpp
@@ -172,9 +172,10 @@ public:
         }
 
         result.appendNumber(HelloCommandReply::kMaxBsonObjectSizeFieldName, BSONObjMaxUserSize);
-        result.appendNumber(HelloCommandReply::kMaxMessageSizeBytesFieldName, MaxMessageSizeBytes);
+        result.appendNumber(HelloCommandReply::kMaxMessageSizeBytesFieldName,
+                            static_cast<long long>(MaxMessageSizeBytes));
         result.appendNumber(HelloCommandReply::kMaxWriteBatchSizeFieldName,
-                            write_ops::kMaxWriteBatchSize);
+                            static_cast<long long>(write_ops::kMaxWriteBatchSize));
         result.appendDate(HelloCommandReply::kLocalTimeFieldName, jsTime());
         result.append(HelloCommandReply::kLogicalSessionTimeoutMinutesFieldName,
                       localLogicalSessionTimeoutMinutes);
